requirements.txt
joblib==1.5.0
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
pillow_heif==0.22.0
scikit-learn==1.6.1
scipy==1.15.3
threadpoolctl==3.6.0
